Create package in ui state model
