+++ presentation/src/main/java/com/pocs/presentation/mapper/PostDetailMapper.kt
@@ -1,7 +1,7 @@
 package com.pocs.presentation.mapper
 
 import com.pocs.domain.model.post.PostDetail
-import com.pocs.presentation.model.PostDetailUiState
+import com.pocs.presentation.model.post.PostDetailUiState
 
 fun PostDetail.toSuccessUiState() = PostDetailUiState.Success(
     id = id,
