+++ presentation/src/main/java/com/pocs/presentation/mapper/PostMapper.kt
@@ -1,7 +1,7 @@
 package com.pocs.presentation.mapper
 
 import com.pocs.domain.model.post.Post
-import com.pocs.presentation.model.PostItemUiState
+import com.pocs.presentation.model.post.PostItemUiState
 
 fun Post.toUiState() = PostItemUiState(
     id = id,
