+++ presentation/src/main/java/com/pocs/presentation/model/post/PostDetailUiState.kt
@@ -1,4 +1,4 @@
-package com.pocs.presentation.model
+package com.pocs.presentation.model.post
 
 import com.pocs.domain.model.post.PostCategory
