+++ presentation/src/main/java/com/pocs/presentation/model/user/UserEditUiState.kt
@@ -1,4 +1,4 @@
-package com.pocs.presentation.model
+package com.pocs.presentation.model.user
 
 data class UserEditUiState(
     val name: String,
