+++ presentation/src/main/java/com/pocs/presentation/model/user/UserUiState.kt
@@ -1,4 +1,4 @@
-package com.pocs.presentation.model
+package com.pocs.presentation.model.user
 
 import androidx.paging.PagingData
