+++ presentation/src/main/java/com/pocs/presentation/view/post/create/PostCreateScreen.kt
@@ -3,7 +3,7 @@ package com.pocs.presentation.view.post.create
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.res.stringResource
 import com.pocs.presentation.R
-import com.pocs.presentation.model.PostCreateUiState
+import com.pocs.presentation.model.post.PostCreateUiState
 import com.pocs.presentation.view.post.edit.PostEditContent
 
 @Composable
