+++ presentation/src/main/java/com/pocs/presentation/view/post/create/PostCreateViewModel.kt
@@ -6,7 +6,7 @@ import androidx.compose.runtime.mutableStateOf
 import androidx.lifecycle.ViewModel
 import com.pocs.domain.model.post.PostCategory
 import com.pocs.domain.usecase.post.AddPostUseCase
-import com.pocs.presentation.model.PostCreateUiState
+import com.pocs.presentation.model.post.PostCreateUiState
 import dagger.hilt.android.lifecycle.HiltViewModel
 import javax.inject.Inject
