+++ presentation/src/main/java/com/pocs/presentation/view/post/edit/PostEditScreen.kt
@@ -15,7 +15,7 @@ import androidx.compose.ui.unit.dp
 import com.pocs.domain.model.post.PostCategory
 import com.pocs.presentation.R
 import com.pocs.presentation.model.BasePostEditUiState
-import com.pocs.presentation.model.PostEditUiState
+import com.pocs.presentation.model.post.PostEditUiState
 import com.pocs.presentation.view.component.RecheckHandler
 import com.pocs.presentation.view.component.appbar.EditContentAppBar
 import kotlinx.coroutines.launch
