+++ presentation/src/main/java/com/pocs/presentation/view/user/UserViewModel.kt
@@ -3,7 +3,7 @@ package com.pocs.presentation.view.user
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.pocs.presentation.mock.mockUsersPagingData
-import com.pocs.presentation.model.UserUiState
+import com.pocs.presentation.model.user.UserUiState
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.asStateFlow
