+++ presentation/src/main/java/com/pocs/presentation/view/user/edit/UserEditViewModel.kt
@@ -5,7 +5,7 @@ import androidx.compose.runtime.State
 import androidx.compose.runtime.mutableStateOf
 import androidx.lifecycle.ViewModel
 import com.pocs.domain.model.user.UserDetail
-import com.pocs.presentation.model.UserEditUiState
+import com.pocs.presentation.model.user.UserEditUiState
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.delay
 import javax.inject.Inject
